Enable VSCode extension APIs for opening notebooks. (#18080)
* Also moved notebook conversion classes to a common folder, and moved several conversion utility methods to a common file.
