+++ src/sql/base/common/locConstants.ts
@@ -48,3 +48,4 @@ export const desktopContributionMiinstallVsix = localize({ key: 'miinstallVsix',
 export const workspaceTrustDescription = localize('workspace.trust.description', "Controls whether or not workspace trust is enabled within Azure Data Studio.");
 export function workspaceTrustEmptyWindowDescription(settingName: string): string { return localize('workspace.trust.emptyWindow.description', "Controls whether or not the empty window is trusted by default within Azure Data Studio. When used with `#{0}#`, you can enable the full functionality of Azure Data Studio without prompting in an empty window.", settingName); }
 export const functionalityNotSupportedError = localize('vscodeFunctionalityNotSupportedError', "This VS Code functionality is not supported in Azure Data Studio.");
+export const invalidArgumentsError = localize('vscodeInvalidArgumentsError', "Invalid arguments");
