+++ src/sql/workbench/api/common/sqlExtHost.protocol.ts
@@ -1007,6 +1007,7 @@ export interface MainThreadNotebookDocumentsAndEditorsShape extends IDisposable 
 	$clearAllOutputs(id: string): Promise<boolean>;
 	$changeKernel(id: string, kernel: azdata.nb.IKernelSpec): Promise<boolean>;
 	$registerNavigationProvider(providerId: string, handle: number);
+	$createNotebookDocument(providerId: string, contents: azdata.nb.INotebookContents): Promise<azdata.nb.NotebookDocument>;
 }
 
 export interface ExtHostExtensionManagementShape {
